python_vector_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
numpy
faiss-cpu
pymongo
sentence-transformers
torch
